requirements.txt
attrs==23.2.0
black==24.3.0
certifi==2024.7.4
cfgv==3.4.0
charset-normalizer==3.3.2
click==8.0.4
darglint==1.8.1
distlib==0.3.8
dlint==0.14.1
dparse==0.6.3
eradicate==2.3.0
filelock==3.15.4
flake8==7.0.0
flake8-comprehensions==3.10.1
flake8-eradicate==1.5.0
Flake8-pyproject==1.2.3
flake8-spellcheck==0.28.0
flake8-typing-imports==1.12.0
identify==2.6.0
idna==3.7
iniconfig==2.0.0
isort==5.8.0
mccabe==0.7.0
mypy==1.8.0
mypy-extensions==1.0.0
nodeenv==1.9.1
packaging==24.1
pathspec==0.12.1
pep8-naming==0.13.2
platformdirs==4.2.2
pluggy==1.5.0
pre-commit==2.20.0
pycodestyle==2.11.1
pyflakes==3.2.0
pytest==8.1.2
pytest-helm-templates==0.0.1a10
pytest-watcher==0.4.2
PyYAML==6.0.1
removestar==1.3.1
requests==2.32.3
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
safety==2.3.4
toml==0.10.2
types-PyYAML==6.0.12.20240724
typing_extensions==4.12.2
urllib3==2.2.2
virtualenv==20.26.3
watchdog==4.0.1
yamllint==1.35.1
